+++ src/main/java/com/example/JavaProject/controller/RecipeController.java
@@ -1,12 +1,8 @@
 package com.example.JavaProject.controller;
 
 import com.example.JavaProject.dto.RecipeDto;
-import com.example.JavaProject.entity.Recipe;
-import com.example.JavaProject.repository.IngredientRepository;
-import com.example.JavaProject.repository.RecipeRepository;
-import com.example.JavaProject.response.LikesCountResponse;
+
 import com.example.JavaProject.response.RecipeResponse;
-import com.example.JavaProject.service.interfaces.AuthenticationService;
 import com.example.JavaProject.service.interfaces.RecipeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -22,7 +18,6 @@ import java.util.List;
 public class RecipeController {
 
     private final RecipeService recipeService;
-    private final AuthenticationService authenticationService;
 
     @GetMapping("/recipes")
     public ResponseEntity<List<RecipeResponse>> getAllRecipe() {
@@ -35,65 +30,18 @@ public class RecipeController {
     }
 
     @PostMapping("/recipes")
-//    @PreAuthorize("isAuthenticated()")
     public ResponseEntity<String> addRecipe(@RequestBody RecipeDto recipeDto) {
         return new ResponseEntity<>(recipeService.addRecipe(recipeDto), HttpStatus.OK);
     }
 
     @PutMapping("/recipes/{id}")
-//    @PreAuthorize("isAuthenticated()")
     public ResponseEntity<String> modifyRecipe(@PathVariable long id, @RequestBody RecipeDto recipeDto) {
         return new ResponseEntity<>(recipeService.modifyRecipe(id, recipeDto), HttpStatus.OK);
     }
-
-    //TODO: brak usówania przepisu <-- dopytaj
 
     @GetMapping("/recipes/search")
     public ResponseEntity<List<RecipeResponse>> findRecipesByIngredient(@RequestParam String ingredientName) {
         List<RecipeResponse> recipes = recipeService.findRecipesByIngredient(ingredientName);
         return new ResponseEntity<>(recipes, HttpStatus.OK);
-    }
-
-
-
-    //FIXME: LIKES:
-    @GetMapping("/recipes/{id}/likes")      //TODO: skąd mam wiedzieć jakie id???
-    public ResponseEntity<LikesCountResponse> getRecipeLikeCount(@PathVariable long id) {
-        return new ResponseEntity<>(recipeService.getRecipeLikes(id), HttpStatus.OK);
-    }
-
-    @PutMapping("/recipes/{id}/likes")
-//    @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<String> addRecipeLike(@PathVariable long id) {
-        return new ResponseEntity<>(recipeService.addLike(id), HttpStatus.OK);
-    }
-
-    @PutMapping("recipes/{id}/unlikes")
-//    @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<String> removeRecipeLike(@PathVariable long id) {
-        return new ResponseEntity<>(recipeService.removeLike(id), HttpStatus.OK);
-    }
-
-
-    //FIXME: FAVOURITES:
-    //FIXME: liczba pojedyncza, czy mnoga? rozważ zmianę endpointów
-    @PutMapping("/recipes/{id}/favourite")
-    @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<String> addRecipeToFavourite(@PathVariable long id) {
-        String message = recipeService.addToFavourite(id);
-        return new ResponseEntity<>(message, HttpStatus.OK);
-    }
-
-    @GetMapping("/recipes/favourites")
-//    @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<List<RecipeResponse>> getFavouriteRecipes() {
-        return new ResponseEntity<>(recipeService.getFavouriteRecipes(), HttpStatus.OK);
-    }
-
-    @DeleteMapping("/recipes/{id}/favourite")
-//    @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<String> deleteFavouriteRecipe(@PathVariable Long id) {
-        String message = recipeService.deleteFavouriteRecipe(id);
-        return new ResponseEntity<>(message, HttpStatus.OK);
     }
 }
